requirements.txt
# DisplayVQA Requirements

# Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0

# Computer Vision and Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine Learning and Metrics
scikit-learn>=1.0.0
matplotlib>=3.4.0

# Model Components
timm>=0.6.0
transformers>=4.15.0

# Data Processing
tqdm>=4.62.0

# Utilities
pathlib2>=2.3.6
argparse>=1.4.0